+++ client/src/proxgui.cpp
@@ -81,6 +81,20 @@ extern "C" void ShowPictureWindow(char *fn) {
     gui->ShowPictureWindow(fn);
 }
 
+extern "C" void ShowBase64PictureWindow(char *b64) {
+    if (!gui) {
+        // Show a notice if X11/XQuartz isn't available
+#if defined(__MACH__) && defined(__APPLE__)
+        PrintAndLogEx(WARNING, "You appear to be on a MacOS device without XQuartz.\nYou may need to install XQuartz (https://www.xquartz.org/) to make the plot work.");
+#else
+        PrintAndLogEx(WARNING, "You appear to be on an environment without an X11 server or without DISPLAY environment variable set.\nPlot may not work until you resolve these issues.");
+#endif
+        return;
+    }
+
+    gui->ShowBase64PictureWindow(b64);
+}
+
 extern "C" void HidePictureWindow(void) {
     if (!gui)
         return;